productos controller implementado filtrado backend

+++ API/Sena/Controllers/Usuarios_control.js
@@ -1,44 +1,28 @@
 import { DBPostgres } from '../../BDPostgres.js'
 //se crea el usuarios segun sun datos
 export const crear_usuarios = async (req, res) => {
-  let { sub, name, email } = req.body
-
-  let id = sub
-  let nombre = name
-
-  console.log(id, nombre, email)
-  //ya esta validado en la base de datos
+  let { id, name, email } = req.body
   //verifica que los datos enviados existan o se rechaza la peticion
-  if (!nombre || !email || !id) {
+  if (!id || !name || !email) {
     return res.status(400).json({
       Error: 'Faltan campos'
     })
   }
+  //verifica que el usuario no exista
+  const existe = await DBPostgres.query(
+    `select * from sena.usuarios where email = $1 `,
+    [email]
+  )
+  if (existe.rows.length) {
+    return res.status(400).json({ error: 'el usuarios ya existe' })
+  }
 
   try {
-    //      1. Verificar si el usuario existe por medio de subconsulta (con parámetros seguros)
-    const usuarioExistente = await DBPostgres.query(
-      ' select exists( SELECT email FROM sena.usuarios WHERE id = $1 and email = $2) ',
-      [id, email]
+    const nuevoUsuario = await DBPostgres.query(
+      `INSERT INTO sena.usuarios (id, nombre, email) VALUES ($1, $2, $3) RETURNING id`,
+      [id, name, email]
     )
-
-    if (usuarioExistente.rows[0].exists) {
-      return res.status(400).json({ Existe: 'El email ya está registrado' })
-    }
-
-    // 2. Insertar usuario (con parámetros seguros)
-    const { rows } = await DBPostgres.query(
-      `INSERT INTO sena.usuarios (id, nombre,email ) 
-       VALUES ($1, $2 ,$3) RETURNING id`,
-      [id, nombre, email]
-    )
-    // 3. Crear carrito y favoritos (en paralelo)
-
-    await DBPostgres.query(`INSERT INTO sena.Carritos (id) VALUES ($1)`, [
-      rows[0].id
-    ])
-
-    return res.status(200).json({ Exito: rows[0].id })
+    return res.status(201).json({ id: nuevoUsuario.rows[0].id })
   } catch (error) {
     return res.status(500).json({ Error: error })
   }
@@ -65,18 +49,51 @@ export const todos_usuarios = async (req, res) => {
 
 export const borrar_usuarios = async (req, res) => {
   const { id } = req.params
+  //verifica que los datos enviados existan o se rechaza la peticion
+  if (!id) {
+    return res.status(400).json({
+      Error: 'Faltan campos'
+    })
+  }
+
   try {
+    //      1. Verificar si el usuario existe por medio de subconsulta (con parámetros seguros)
+    const usuarioExistente = await DBPostgres.query(
+      ' select exists( SELECT id FROM sena.usuarios WHERE id = $1) ',
+      [id]
+    )
+    if (!usuarioExistente.rows[0].exists) {
+      return res.status(400).json({ Existe: 'El usuario no existe' })
+    }
+    // 2. Borrar usuario (con parámetros seguros)
     const { rows } = await DBPostgres.query(
       'delete from sena.usuarios where id = $1 returning id',
       [id]
     )
+    // 3. Borrar carrito y favoritos (en paralelo)
+    await DBPostgres.query(
+      `delete from sena.carritos where id = $1 returning id`,
+      [id]
+    )
+    await DBPostgres.query(
+      `delete from sena.favoritos where id = $1 returning id`,
+      [id]
+    )
+    // 4. Verificar si el usuario fue
+    // borrado correctamente
 
-    if (!rows.length) {
+    const usuarioBorrado = rows.length > 0
+
+    if (!usuarioBorrado) {
       return res.status(404).json({ error: 'usuario no encontrado' })
     }
 
-    return res.status(200).json({ Borrado: rows })
+    return res
+      .status(200)
+      .json({ Borrado: rows, mensaje: 'Usuario borrado correctamente' })
   } catch (error) {
+    // Manejo de errores
+
     return res.status(500).json({
       errores: error
     })
@@ -86,6 +103,13 @@ export const borrar_usuarios = async (req, res) => {
 export const Actualizar_usuarios = async (req, res) => {
   const { nombre, edad, email, claveacceso } = req.body
   const { id } = req.params
+  //verifica que los datos enviados existan o se rechaza la peticion
+  if (!nombre || !edad || !email || !claveacceso || !id) {
+    return res.status(400).json({
+      Error: 'Faltan campos'
+    })
+  }
+
   try {
     const { rows } = await DBPostgres.query(
       `UPDATE sena.usuarios 
@@ -205,16 +229,6 @@ export const borrar_usuarios_admin = async (req, res) => {
     }
 
     return res.status(200).json({ Borrado: rows })
-  } catch (error) {
-    return res.status(500).json({
-      errores: error
-    })
-  }
-}
-
-export const tost = async (req, res) => {
-  try {
-    return res.status(200).json({ exito: 'exito' })
   } catch (error) {
     return res.status(500).json({
       errores: error
